pyproject.toml
[tool.poetry]
name = "gg-bot"
version = "0.1.0"
description = ""
authors = ["Boris Shishov <borisshishov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.0"
PyYAML = "^5.3.1"
Jinja2 = "^2.11.2"
requests = "^2.25.0"
PyStemmer = "^2.0.1"
pickledb = "^0.9.2"
tokema = "^0.0.2"
numpy = "^1.20.2"
gspread = "^3.7.0"


[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
